pyproject.toml
[tool.poetry]
name = "speak2me-fitness-monorepo"
version = "0.1.0"
description = ""
authors = ["stazman <cadistasio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
pre-commit = "^3.8.0"
flake8 = "^7.1.1"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
motor = "^3.6.0"
beanie = "^1.27.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
coverage = "^7.6.1"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
